Update all assets to the corresponding versions on the cypress-202002.00 branches. Note: This has not been divided into commit per asset because this is expected to be an interim state and versions of multiple assets are not known on the cypress-202002.00 branch. These assets will be updated/submoduled to released versions from cypress github. This commit also contains some fixes that were required to get these version of the assets working on the 202007.00 branch.

+++ vendors/cypress/MTB/psoc6/psoc6csp/hal/COMPONENT_PSOC6HAL/source/cyhal_crypto_common.c
@@ -57,10 +57,10 @@ static uint16_t cyhal_crypto_features_max_val[CRYPTO_FEATURES_NUM] = {CRYPTO_FEA
 
 static bool cy_crypto_enabled(uint32_t instance_count)
 {
-    uint8_t reserved = (cyhal_crypto_features[instance_count][CYHAL_CRYPTO_CRC]  |
-                        cyhal_crypto_features[instance_count][CYHAL_CRYPTO_TRNG] |
-                        cyhal_crypto_features[instance_count][CYHAL_CRYPTO_VU]   |
-                        cyhal_crypto_features[instance_count][CYHAL_CRYPTO_COMMON]);
+    uint8_t reserved = (uint8_t)(cyhal_crypto_features[instance_count][CYHAL_CRYPTO_CRC]  |
+                                 cyhal_crypto_features[instance_count][CYHAL_CRYPTO_TRNG] |
+                                 cyhal_crypto_features[instance_count][CYHAL_CRYPTO_VU]   |
+                                 cyhal_crypto_features[instance_count][CYHAL_CRYPTO_COMMON]);
 
     return (reserved != 0);
 }
@@ -74,7 +74,7 @@ cy_rslt_t cyhal_crypto_reserve(CRYPTO_Type** base, cyhal_resource_inst_t *resour
         if (cyhal_crypto_features[i][feature] < cyhal_crypto_features_max_val[feature])
         {
             resource->type = CYHAL_RSC_CRYPTO;
-            resource->block_num = i;
+            resource->block_num = (uint8_t)i;
             resource->channel_num = 0;
             *base = CYHAL_CRYPTO_BASE_ADDRESSES[i];
